requirements.txt
certifi==2023.7.22
charset-normalizer==2.1.1
idna==3.4
Pillow==9.3.0
python-dotenv==1.0.0
requests==2.28.1
urllib3==1.26.17
